refactoring

+++ TelegramBotServer/Repository/DbEventRepository.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
+using System.Data.Common;
 using System.Linq;
 using System.Threading.Tasks;
 using TelegramBotServer.DatabaseContext;
@@ -25,46 +26,53 @@ namespace TelegramBotServer.Repository
         {
             var dbContext = _scopeFactory.CreateScope()
                 .ServiceProvider.GetRequiredService<UsersContext>();
+            if (dbContext is null || dbContext.Events is null)
+                throw new Exception("Database context is null!");
+            else
+            {
+                var validator = new EventValidator(dbContext.Events);
+                if (!validator.Validate(newEvent))
+                    throw new ArgumentException("Invalid event");
 
-            var validator = new EventValidator(dbContext.Events);
-            if (!validator.Validate(newEvent))
-                throw new ArgumentException("Invalid event");
+                dbContext.Events.Add(newEvent);
+                dbContext.SaveChanges();
 
-            dbContext.Events.Add(newEvent);
-            dbContext.SaveChanges();
-
-            _logger.LogInformation($"Add Event with id {newEvent.Id}");
-
-            return newEvent.Id;
+                _logger.LogInformation($"Add Event with id {newEvent.Id}");
+                return newEvent.Id;
+            }
         }
 
         public void ChangeEvent(int id, Event newEvent)
         {
             var dbContext = _scopeFactory.CreateScope()
                 .ServiceProvider.GetRequiredService<UsersContext>();
+            if (dbContext is null || dbContext.Events is null)
+                throw new Exception("Database context is null!");
+            else
+            {
+                var validator = new EventValidator(dbContext.Events);
+                if (!validator.Validate(newEvent))
+                    throw new ArgumentException("Invalid event");
 
-            var validator = new EventValidator(dbContext.Events);
-            if (!validator.Validate(newEvent))
-                throw new ArgumentException("Invalid event");
+                var chEvent = dbContext.Events.FirstOrDefault(e => e.Id == id);
 
-            var chEvent = dbContext.Events.FirstOrDefault(e => e.Id == id);
+                if (chEvent is null)
+                    return;
 
-            if (chEvent is null)
-                return;
-
-            dbContext.Update(chEvent).CurrentValues.SetValues(newEvent);
-            dbContext.SaveChanges();
+                dbContext.Update(chEvent).CurrentValues.SetValues(newEvent);
+                dbContext.SaveChanges();
+            }
         }
 
-        public Event GetEvent(int id)
+        public Event? GetEvent(int id)
         {
             var dbContext = _scopeFactory.CreateScope()
                 .ServiceProvider.GetRequiredService<UsersContext>();
 
-            return dbContext.Events.Find(id);
+            return dbContext?.Events?.Find(id);
         }
 
-        public IEnumerable<Event> GetEvents()
+        public IEnumerable<Event>? GetEvents()
         {
             var dbContext = _scopeFactory.CreateScope()
                 .ServiceProvider.GetRequiredService<UsersContext>();
@@ -76,18 +84,22 @@ namespace TelegramBotServer.Repository
         {
             var dbContext = _scopeFactory.CreateScope()
                 .ServiceProvider.GetRequiredService<UsersContext>();
+            if (dbContext is null || dbContext.Events is null)
+                throw new Exception("Database context is null!");
+            else
+            {
+                var delEvent = dbContext.Events.FirstOrDefault(e => e.Id == id);
 
-            var delEvent = dbContext.Events.FirstOrDefault(e => e.Id == id);
+                if (delEvent is null)
+                    return false;
 
-            if (delEvent is null)
-                return false;
+                dbContext.Events.Remove(delEvent);
+                dbContext.SaveChanges();
 
-            dbContext.Events.Remove(delEvent);
-            dbContext.SaveChanges();
+                _logger.LogInformation($"Remove Event with id {delEvent.Id}");
 
-            _logger.LogInformation($"Remove Event with id {delEvent.Id}");
-
-            return true;
+                return true;
+            }
         }
     }
 }
